refactor: split module, export "v" namespace
