+++ packages/x6/package.json
@@ -43,7 +43,8 @@
     "build": "run-s build:version build:dev",
     "prebuild": "run-s lint clean",
     "test": "jest --coverage",
-    "test-debug": "DEBUG_MODE=1 jest",
+    "test:watch": "jest --watch --coverage",
+    "test:debug": "DEBUG_MODE=1 jest",
     "coveralls": "cat ./test/coverage/lcov.info | coveralls",
     "pretest": "run-p clean:coverage",
     "prepare": "run-s build:version test build",
