+++ packages/x6/src/v/matrix.test.ts
@@ -1,0 +1,269 @@
+import { v } from './v'
+import { Vectorizer } from './vectorizer'
+
+describe('v', () => {
+  const fixture = document.createElement('div')
+  const svgContainer = v('svg').node
+  fixture.appendChild(svgContainer)
+  document.body.appendChild(fixture)
+
+  afterAll(() => {
+    fixture.parentNode?.removeChild(fixture)
+  })
+
+  describe('#transformPoint', () => {
+    const p = { x: 1, y: 2 }
+    const group = v('g')
+    const node = group.node as SVGGraphicsElement
+    svgContainer.appendChild(group.node)
+
+    it('should return the point unchanged when transformed without transformation', () => {
+      const t = v.transformPoint(p, node.getCTM()!)
+      expect(t.equals(p)).toBe(true)
+    })
+    it('should returns correct point when transformed with scale transformation', () => {
+      group.scale(2, 3)
+      const t = v.transformPoint(p, node.getCTM()!)
+      expect(t).toEqual({ x: 2, y: 6 })
+    })
+    it('should returns correct point when transformed with rotate transformation', () => {
+      group.attr('transform', 'rotate(90)')
+      const t = v.transformPoint(p, node.getCTM()!)
+      expect(t).toEqual({ x: -2, y: 1 })
+    })
+  })
+
+  describe('#getTransformToElement', () => {
+    it('rotate', () => {
+      const normalizeFloat = (value: number) => {
+        const temp = value * 100
+        return temp > 0 ? Math.floor(temp) : Math.ceil(temp)
+      }
+
+      const container = v(svgContainer)
+      const group = v('g')
+      const rect = v('rect')
+
+      container.append(group)
+      container.append(rect)
+
+      rect.rotate(45)
+
+      const matrix = group.getTransformToElement(rect.node)
+      expect({
+        a: normalizeFloat(matrix.a),
+        b: normalizeFloat(matrix.b),
+        c: normalizeFloat(matrix.c),
+        d: normalizeFloat(matrix.d),
+        e: normalizeFloat(matrix.e),
+        f: normalizeFloat(matrix.f),
+      }).toEqual({
+        a: normalizeFloat(0.7071067811865476),
+        b: normalizeFloat(-0.7071067811865475),
+        c: normalizeFloat(0.7071067811865475),
+        d: normalizeFloat(0.7071067811865476),
+        e: normalizeFloat(0),
+        f: normalizeFloat(0),
+      })
+
+      group.remove()
+      rect.remove()
+    })
+
+    it('translate', () => {
+      const container = v(svgContainer)
+      const group = v('g')
+      const rect = v('rect')
+
+      container.append(group)
+      container.append(rect)
+
+      rect.translate(10, 10)
+
+      const matrix = group.getTransformToElement(rect.node)
+      expect({
+        a: matrix.a,
+        b: matrix.b,
+        c: matrix.c,
+        d: matrix.d,
+        e: matrix.e,
+        f: matrix.f,
+      }).toEqual({
+        a: 1,
+        b: 0,
+        c: 0,
+        d: 1,
+        e: -10,
+        f: -10,
+      })
+
+      group.remove()
+      rect.remove()
+    })
+  })
+
+  describe('#parseTransformString', () => {
+    it('should parse scale, rotate, translate', () => {
+      const parsed = v.parseTransformString(
+        'scale(3) rotate(6) translate(9) xxx(11)',
+      )
+
+      expect(parsed.scale).toEqual({ sx: 3, sy: 3 })
+      expect(parsed.rotate).toEqual({
+        angle: 6,
+        cx: undefined,
+        cy: undefined,
+      })
+      expect(parsed.translate).toEqual({ tx: 9, ty: 0 })
+    })
+  })
+
+  describe('#transformStringToMatrix', () => {
+    let svgTestGroup: Vectorizer
+
+    beforeEach(() => {
+      svgTestGroup = v('g')
+      svgContainer.appendChild(svgTestGroup.node)
+    })
+
+    afterEach(() => {
+      svgTestGroup.remove()
+    })
+
+    const arr = [
+      '',
+      'scale(2)',
+      'scale(2,3)',
+      'scale(2.5,3.1)',
+      'translate(10, 10)',
+      'translate(10,10)',
+      'translate(10.2,11.6)',
+      'rotate(10)',
+      'rotate(10,100,100)',
+      'skewX(40)',
+      'skewY(60)',
+      'scale(2,2) matrix(1 0 0 1 10 10)',
+      'matrix(1 0 0 1 10 10) scale(2,2)',
+      'rotate(10,100,100) matrix(1 0 0 1 10 10) scale(2,2) translate(10,20)',
+    ]
+
+    arr.forEach(transformString => {
+      it(`should convert "${transformString}" to matrix`, () => {
+        svgTestGroup.attr('transform', transformString)
+        expect(v.transformStringToMatrix(transformString)).toEqual(
+          (svgTestGroup.node as SVGGraphicsElement).getCTM() as any,
+        )
+      })
+    })
+  })
+
+  describe('#matrixToTransformString', () => {
+    it('should return correct transformation string', () => {
+      expect(v.matrixToTransformString()).toEqual('matrix(1,0,0,1,0,0)')
+      expect(v.matrixToTransformString({ a: 2, d: 2 })).toEqual(
+        'matrix(2,0,0,2,0,0)',
+      )
+
+      expect(
+        v.matrixToTransformString({ a: 1, b: 2, c: 3, d: 4, e: 5, f: 6 }),
+      ).toEqual('matrix(1,2,3,4,5,6)')
+
+      expect(
+        v.matrixToTransformString(
+          v.createSVGMatrix({ a: 1, b: 2, c: 3, d: 4, e: 5, f: 6 }),
+        ),
+      ).toEqual('matrix(1,2,3,4,5,6)')
+      expect(
+        v.matrixToTransformString({ a: 0, b: 1, c: 1, d: 0, e: 0, f: 0 }),
+      ).toEqual('matrix(0,1,1,0,0,0)')
+    })
+  })
+
+  describe('#matrixTo[Transformation]', () => {
+    function roundObject(obj: any) {
+      for (const i in obj) {
+        if (obj.hasOwnProperty(i)) {
+          obj[i] = Math.round(obj[i])
+        }
+      }
+      return obj
+    }
+
+    it('should convert matrix to rotation metadata', () => {
+      let angle
+      angle = v.matrixToRotate(v.createSVGMatrix().rotate(45))
+      expect(roundObject(angle)).toEqual({ angle: 45 })
+
+      angle = v.matrixToRotate(
+        v
+          .createSVGMatrix()
+          .translate(50, 50)
+          .rotate(15),
+      )
+      expect(roundObject(angle)).toEqual({ angle: 15 })
+
+      angle = v.matrixToRotate(
+        v
+          .createSVGMatrix()
+          .translate(50, 50)
+          .rotate(60)
+          .scale(2),
+      )
+      expect(roundObject(angle)).toEqual({ angle: 60 })
+
+      angle = v.matrixToRotate(
+        v
+          .createSVGMatrix()
+          .rotate(60)
+          .rotate(60),
+      )
+      expect(roundObject(angle)).toEqual({ angle: 120 })
+    })
+
+    it('should convert matrix to translation medata', () => {
+      let translate
+      translate = v.matrixToTranslate(v.createSVGMatrix().translate(10, 20))
+      expect(roundObject(translate)).toEqual({ tx: 10, ty: 20 })
+
+      translate = v.matrixToTranslate(
+        v
+          .createSVGMatrix()
+          .translate(10, 20)
+          .rotate(10, 20)
+          .scale(2),
+      )
+      expect(roundObject(translate)).toEqual({ tx: 10, ty: 20 })
+
+      translate = v.matrixToTranslate(
+        v
+          .createSVGMatrix()
+          .translate(10, 20)
+          .translate(30, 40),
+      )
+      expect(roundObject(translate)).toEqual({ tx: 40, ty: 60 })
+    })
+
+    it('should convert matrix to scaling metadata', () => {
+      let scale
+      scale = v.matrixToScale(v.createSVGMatrix().scale(2))
+      expect(roundObject(scale)).toEqual({ sx: 2, sy: 2 })
+
+      scale = v.matrixToScale(
+        v
+          .createSVGMatrix()
+          .translate(15, 15)
+          .scaleNonUniform(2, 3)
+          .rotate(10, 20),
+      )
+      expect(roundObject(scale)).toEqual({ sx: 2, sy: 3 })
+
+      scale = v.matrixToScale(
+        v
+          .createSVGMatrix()
+          .scale(2, 2)
+          .scale(3, 3),
+      )
+      expect(roundObject(scale)).toEqual({ sx: 6, sy: 6 })
+    })
+  })
+})
